littlelemon/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djoser>=2.2.0,<3.0.0
django-cors-headers>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
djangorestframework-simplejwt>=5.2.2,<6.0.0
python-dotenv>=1.0.0,<2.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
pytest>=7.0.0,<8.0.0
pytest-django>=4.5.0,<5.0.0
